# Conflicts:
#	src/records/User.php
#	src/templates/assets/index.html
#	src/templates/users/_edit.html

+++ src/services/Fields.php
@@ -195,38 +195,25 @@ class Fields extends Component
     /**
      * Returns all field groups.
      *
-     * @param string|null $indexBy The attribute to index the field groups by
-     *
      * @return FieldGroup[] The field groups
      */
-    public function getAllGroups($indexBy = null)
+    public function getAllGroups()
     {
-        if (!$this->_fetchedAllGroups) {
-            $this->_groupsById = [];
-
-            $results = $this->_createGroupQuery()->all();
-
-            foreach ($results as $result) {
-                $group = new FieldGroup($result);
-                $this->_groupsById[$group->id] = $group;
-            }
-
-            $this->_fetchedAllGroups = true;
+        if ($this->_fetchedAllGroups) {
+            return array_values($this->_groupsById);
         }
 
-        if ($indexBy == 'id') {
-            $groups = $this->_groupsById;
-        } else if (!$indexBy) {
-            $groups = array_values($this->_groupsById);
-        } else {
-            $groups = [];
+        $this->_groupsById = [];
+        $results = $this->_createGroupQuery()->all();
 
-            foreach ($this->_groupsById as $group) {
-                $groups[$group->$indexBy] = $group;
-            }
+        foreach ($results as $result) {
+            $group = new FieldGroup($result);
+            $this->_groupsById[$group->id] = $group;
         }
 
-        return $groups;
+        $this->_fetchedAllGroups = true;
+
+        return array_values($this->_groupsById);
     }
 
     /**
@@ -238,21 +225,23 @@ class Fields extends Component
      */
     public function getGroupById($groupId)
     {
-        if (!isset($this->_groupsById) || !array_key_exists($groupId, $this->_groupsById)) {
-            $result = $this->_createGroupQuery()
-                ->where(['id' => $groupId])
-                ->one();
-
-            if ($result) {
-                $group = new FieldGroup($result);
-            } else {
-                $group = null;
-            }
-
-            $this->_groupsById[$groupId] = $group;
+        if ($this->_groupsById !== null && array_key_exists($groupId, $this->_groupsById)) {
+            return $this->_groupsById[$groupId];
         }
 
-        return $this->_groupsById[$groupId];
+        if ($this->_fetchedAllGroups) {
+            return null;
+        }
+
+        $result = $this->_createGroupQuery()
+            ->where(['id' => $groupId])
+            ->one();
+
+        if (!$result) {
+            return $this->_groupsById[$groupId] = null;
+        }
+
+        return $this->_groupsById[$groupId] = new FieldGroup($result);
     }
 
     /**
@@ -411,7 +400,8 @@ class Fields extends Component
     {
         $fieldTypes = [];
 
-        foreach (static::getAllFieldTypes() as $fieldType) {
+        foreach ($this->getAllFieldTypes() as $fieldType) {
+            /** @var Field|string $fieldType */
             if ($fieldType::hasContentColumn()) {
                 $fieldTypes[] = $fieldType;
             }
@@ -450,21 +440,21 @@ class Fields extends Component
     /**
      * Returns all fields within a field context(s).
      *
-     * @param string|null          $indexBy The field property to index the resulting fields by
      * @param string|string[]|null $context The field context(s) to fetch fields from. Defaults to {@link ContentService::$fieldContext}.
      *
      * @return FieldInterface[] The fields
      */
-    public function getAllFields($indexBy = null, $context = null)
+    public function getAllFields($context = null)
     {
         if ($context === null) {
             $context = [Craft::$app->getContent()->fieldContext];
         } else if (!is_array($context)) {
-            $context = [$context];
+            $context = (array)$context;
         }
 
         $missingContexts = [];
 
+        /** @noinspection ForeachSourceInspection - FP */
         foreach ($context as $c) {
             if (!isset($this->_allFieldsInContext[$c])) {
                 $missingContexts[] = $c;
@@ -489,13 +479,10 @@ class Fields extends Component
 
         $fields = [];
 
+        /** @noinspection ForeachSourceInspection - FP */
         foreach ($context as $c) {
-            if (!$indexBy) {
-                $fields = array_merge($fields, $this->_allFieldsInContext[$c]);
-            } else {
-                foreach ($this->_allFieldsInContext[$c] as $field) {
-                    $fields[$field->$indexBy] = $field;
-                }
+            foreach ($this->_allFieldsInContext[$c] as $field) {
+                $fields[] = $field;
             }
         }
 
@@ -533,23 +520,24 @@ class Fields extends Component
      */
     public function getFieldById($fieldId)
     {
-        if (!isset($this->_fieldsById) || !array_key_exists($fieldId, $this->_fieldsById)) {
-            $result = $this->_createFieldQuery()
-                ->where(['fields.id' => $fieldId])
-                ->one();
-
-            if ($result) {
-                /** @var Field $field */
-                $field = $this->createField($result);
-
-                $this->_fieldsById[$fieldId] = $field;
-                $this->_fieldsByContextAndHandle[$field->context][$field->handle] = $field;
-            } else {
-                return null;
-            }
+        if ($this->_fieldsById !== null && array_key_exists($fieldId, $this->_fieldsById)) {
+            return $this->_fieldsById[$fieldId];
         }
 
-        return $this->_fieldsById[$fieldId];
+        $result = $this->_createFieldQuery()
+            ->where(['fields.id' => $fieldId])
+            ->one();
+
+        if (!$result) {
+            return $this->_fieldsById[$fieldId] = null;
+        }
+
+        /** @var Field $field */
+        $field = $this->createField($result);
+        $this->_fieldsById[$fieldId] = $field;
+        $this->_fieldsByContextAndHandle[$field->context][$field->handle] = $field;
+
+        return $field;
     }
 
     /**
@@ -601,7 +589,7 @@ class Fields extends Component
             $context = Craft::$app->getContent()->fieldContext;
         }
 
-        if (!isset($this->_allFieldHandlesByContext)) {
+        if ($this->_allFieldHandlesByContext === null) {
             $this->_allFieldHandlesByContext = [];
 
             $results = (new Query())
@@ -614,18 +602,17 @@ class Fields extends Component
             }
         }
 
-        return (isset($this->_allFieldHandlesByContext[$context]) && in_array($handle, $this->_allFieldHandlesByContext[$context]));
+        return (isset($this->_allFieldHandlesByContext[$context]) && in_array($handle, $this->_allFieldHandlesByContext[$context], true));
     }
 
     /**
      * Returns all the fields in a given group.
      *
-     * @param integer     $groupId The field group’s ID
-     * @param string|null $indexBy The attribute to index the fields by
+     * @param integer $groupId The field group’s ID
      *
      * @return FieldInterface[] The fields
      */
-    public function getFieldsByGroupId($groupId, $indexBy = null)
+    public function getFieldsByGroupId($groupId)
     {
         $results = $this->_createFieldQuery()
             ->where(['fields.groupId' => $groupId])
@@ -634,13 +621,7 @@ class Fields extends Component
         $fields = [];
 
         foreach ($results as $result) {
-            $field = $this->createField($result);
-
-            if ($indexBy) {
-                $fields[$field->$indexBy] = $field;
-            } else {
-                $fields[] = $field;
-            }
+            $fields[] = $this->createField($result);
         }
 
         return $fields;
@@ -650,11 +631,10 @@ class Fields extends Component
      * Returns all of the fields used by a given element type.
      *
      * @param ElementInterface|string $elementType
-     * @param string|null             $indexBy
      *
      * @return FieldInterface[] The fields
      */
-    public function getFieldsByElementType($elementType, $indexBy = null)
+    public function getFieldsByElementType($elementType)
     {
         $results = $this->_createFieldQuery()
             ->innerJoin('{{%fieldlayoutfields}} flf', '[[flf.fieldId]] = [[fields.id]]')
@@ -665,13 +645,7 @@ class Fields extends Component
         $fields = [];
 
         foreach ($results as $result) {
-            $field = $this->createField($result);
-
-            if ($indexBy) {
-                $fields[$field->$indexBy] = $field;
-            } else {
-                $fields[] = $field;
-            }
+            $fields[] = $this->createField($result);
         }
 
         return $fields;
@@ -752,12 +726,14 @@ class Fields extends Component
                 }
             } else {
                 // Did the old field have a column we need to remove?
-                if (!$isNewField) {
-                    if ($fieldRecord->getOldHandle() && Craft::$app->getDb()->columnExists($contentTable, $oldColumnName)) {
-                        Craft::$app->getDb()->createCommand()
-                            ->dropColumn($contentTable, $oldColumnName)
-                            ->execute();
-                    }
+                if (
+                    !$isNewField &&
+                    $fieldRecord->getOldHandle() &&
+                    Craft::$app->getDb()->columnExists($contentTable, $oldColumnName)
+                ) {
+                    Craft::$app->getDb()->createCommand()
+                        ->dropColumn($contentTable, $oldColumnName)
+                        ->execute();
                 }
 
                 // Fields without a content column don't get translated
@@ -789,7 +765,7 @@ class Fields extends Component
                 if (
                     isset($this->_allFieldHandlesByContext[$field->context]) &&
                     $field->oldHandle != $field->handle &&
-                    ($oldHandleIndex = array_search($field->oldHandle, $this->_allFieldHandlesByContext[$field->context])) !== false
+                    ($oldHandleIndex = array_search($field->oldHandle, $this->_allFieldHandlesByContext[$field->context], true)) !== false
                 ) {
                     array_splice($this->_allFieldHandlesByContext[$field->context], $oldHandleIndex, 1);
                 }
@@ -799,12 +775,11 @@ class Fields extends Component
             $this->_fieldsById[$field->id] = $field;
             $this->_fieldsByContextAndHandle[$field->context][$field->handle] = $field;
 
-            if (isset($this->_allFieldHandlesByContext)) {
+            if ($this->_allFieldHandlesByContext !== null) {
                 $this->_allFieldHandlesByContext[$field->context][] = $field->handle;
             }
 
-            unset($this->_allFieldsInContext[$field->context]);
-            unset($this->_fieldsWithContent[$field->context]);
+            unset($this->_allFieldsInContext[$field->context], $this->_fieldsWithContent[$field->context]);
 
             $field->afterSave($isNewField);
 
@@ -919,21 +894,19 @@ class Fields extends Component
      */
     public function getLayoutById($layoutId)
     {
-        if (!isset($this->_layoutsById) || !array_key_exists($layoutId, $this->_layoutsById)) {
-            $result = $this->_createLayoutQuery()
-                ->where(['id' => $layoutId])
-                ->one();
-
-            if ($result) {
-                $layout = new FieldLayout($result);
-            } else {
-                $layout = null;
-            }
-
-            $this->_layoutsById[$layoutId] = $layout;
+        if ($this->_layoutsById !== null && array_key_exists($layoutId, $this->_layoutsById)) {
+            return $this->_layoutsById[$layoutId];
         }
 
-        return $this->_layoutsById[$layoutId];
+        $result = $this->_createLayoutQuery()
+            ->where(['id' => $layoutId])
+            ->one();
+
+        if (!$result) {
+            return $this->_layoutsById[$layoutId] = null;
+        }
+
+        return $this->_layoutsById[$layoutId] = new FieldLayout($result);
     }
 
     /**
@@ -945,27 +918,24 @@ class Fields extends Component
      */
     public function getLayoutByType($type)
     {
-        if (!isset($this->_layoutsByType) || !array_key_exists($type, $this->_layoutsByType)) {
-            $result = $this->_createLayoutQuery()
-                ->where(['type' => $type])
-                ->one();
-
-            if ($result) {
-                $id = $result['id'];
-
-                if (!isset($this->_layoutsById[$id])) {
-                    $this->_layoutsById[$id] = new FieldLayout($result);
-                }
-
-                $layout = $this->_layoutsById[$id];
-            } else {
-                $layout = new FieldLayout();
-            }
-
-            $this->_layoutsByType[$type] = $layout;
+        if ($this->_layoutsByType !== null && array_key_exists($type, $this->_layoutsByType)) {
+            return $this->_layoutsByType[$type];
         }
 
-        return $this->_layoutsByType[$type];
+        $result = $this->_createLayoutQuery()
+            ->where(['type' => $type])
+            ->one();
+
+        if (!$result) {
+            return $this->_layoutsByType[$type] = new FieldLayout();
+        }
+
+        $id = $result['id'];
+        if (!isset($this->_layoutsById[$id])) {
+            $this->_layoutsById[$id] = new FieldLayout($result);
+        }
+
+        return $this->_layoutsByType[$type] = $this->_layoutsById[$id];
     }
 
     /**
@@ -1043,7 +1013,7 @@ class Fields extends Component
      *
      * @return FieldLayout The field layout
      */
-    public function assembleLayout($postedFieldLayout, $requiredFields = [])
+    public function assembleLayout($postedFieldLayout, array $requiredFields = [])
     {
         $tabs = [];
         $fields = [];
@@ -1054,7 +1024,9 @@ class Fields extends Component
         $allFieldIds = [];
 
         foreach ($postedFieldLayout as $fieldIds) {
-            $allFieldIds = array_merge($allFieldIds, $fieldIds);
+            foreach ($fieldIds as $fieldId) {
+                $allFieldIds[] = $fieldId;
+            }
         }
 
         if ($allFieldIds) {
@@ -1079,7 +1051,7 @@ class Fields extends Component
                 }
 
                 $field = $allFieldsById[$fieldId];
-                $field->required = in_array($fieldId, $requiredFields);
+                $field->required = in_array($fieldId, $requiredFields, false);
                 $field->sortOrder = ($fieldSortOrder + 1);
 
                 $fields[] = $field;
@@ -1176,11 +1148,7 @@ class Fields extends Component
             return false;
         }
 
-        if (!is_array($layoutId)) {
-            $layoutId = [$layoutId];
-        }
-
-        foreach ($layoutId as $thisLayoutId) {
+        foreach ((array)$layoutId as $thisLayoutId) {
             $layout = $this->getLayoutById($thisLayoutId);
 
             if ($layout) {
@@ -1344,19 +1312,19 @@ class Fields extends Component
     private function _getFieldRecord(FieldInterface $field)
     {
         /** @var Field $field */
-        if (!$field->getIsNew()) {
-            if (!isset($this->_fieldRecordsById) || !array_key_exists($field->id, $this->_fieldRecordsById)) {
-                $this->_fieldRecordsById[$field->id] = FieldRecord::findOne($field->id);
+        if ($field->getIsNew()) {
+            return new FieldRecord();
+        }
 
-                if (!$this->_fieldRecordsById[$field->id]) {
-                    throw new FieldNotFoundException("No field exists with the ID '{$field->id}'");
-                }
-            }
-
+        if ($this->_fieldRecordsById !== null && array_key_exists($field->id, $this->_fieldRecordsById)) {
             return $this->_fieldRecordsById[$field->id];
         }
 
-        return new FieldRecord();
+        if (($this->_fieldRecordsById[$field->id] = FieldRecord::findOne($field->id)) === null) {
+            throw new FieldNotFoundException('Invalid field ID: '.$field->id);
+        }
+
+        return $this->_fieldRecordsById[$field->id];
     }
 
     /**
